Move test modules to separate files so cargo can skip building them
